Remove hardcoded names in ponies23.list
The ponies23.list file was used to make sure we only use smaller ponies
for motd, so we don't take up too much screen space.

The list of small enough ponies is now generated each time so we can
also use newly added ponies. And the limit for pony size is now up to 27
characters high.